+++ app/controllers/rails_audit_log/application_controller.rb
@@ -1,5 +1,7 @@
 module RailsAuditLog
   class ApplicationController < ActionController::Base
+    include Pagy::Method
+
     protect_from_forgery with: :exception
   end
 end